+++ packages/gestalt/src/Button.test.js
@@ -1,7 +1,6 @@
 // @flow
 import React from 'react';
 import { create } from 'react-test-renderer';
-import { shallow } from 'enzyme';
 import Button from './Button.js';
 
 test('Button with custom color', () => {
@@ -16,11 +15,4 @@ test('Button with custom text color', () => {
     <Button color="white" textColor="blue" text="Hello World" />
   ).toJSON();
   expect(tree).toMatchSnapshot();
-});
-
-test('Button handles click', () => {
-  const mockOnClick = jest.fn();
-  const wrapper = shallow(<Button text="Text" onClick={mockOnClick} />);
-  wrapper.find('button').simulate('click');
-  expect(mockOnClick).toHaveBeenCalled();
 });